It's sad to see computer science go :(
This code was my response for question 9(f) of ocr computer science paper 2, sat on the 21st May 2024 in the afternoon. This was what I wrote on the actual paper apart from the brackets on the .upper string manipulation which I completely forgot about, woops.
This therefore marks the end of GCSE computer science. 
It was a nice time to be alive.